+++ src/app/components/solution.tsx
@@ -1,14 +1,19 @@
 import { useTranslation } from "react-i18next";
-import { Solution, useSettings } from "../store";
+import { ElementType, LineStyle, PointStyle, Solution, useDB, useSettings, useStore } from "../store";
 import { StepComponent } from "./step";
 import { DIAGRAM_ZOOM, Diagram } from "./svg/diagram";
 import { useRef } from "react";
 import { Exact } from "./exact";
+import { resetWorker } from "../bridge";
 
 interface SolutionComponentProps {
 	data: Solution;
 	show: boolean;
 	onSelect: () => void;
+}
+
+interface AddButtonProps {
+	data: Solution;
 }
 
 const DIAGRAM_PADDING = 0.18;
@@ -34,6 +39,58 @@ export function formatSolution(data: Solution, precision: number): string {
 	return `(${data.solution[0].toFixed(precision)}, ${text})`;
 }
 
+/**
+ * Function that runs when the user accepts a specific solution.
+ * Adds the creases from the last diagram of the accepted solution and recreates the database.
+ */
+function addAsExistingCreases(data: Solution) {
+	// Add the creases from the last diagram of the accepted solution to the store (to render on the preview paper)
+	const { existingMarks, existingLines, existingRefs } = useStore.getState();
+	for(let i = 0; i < data.steps.length; i++) {
+		let step = data.steps[i];
+		const diag = data.diagrams[i];
+		if(step.axiom != 0) { // line
+			const lookFor = step.pinch ? LineStyle.pinch : LineStyle.valley;
+			for(const el of diag) {
+				if(el && el.type == ElementType.line && el.style == lookFor) {
+					el.style = LineStyle.crease;
+					existingRefs.push(el);
+					if(!step.pinch) existingLines.push([el.from, el.to]);
+					break;
+				}
+			}
+			if(step.intersection) step = step.intersection; // continue with the next part
+		}
+		if(step.axiom == 0) { // point
+			for(const el of diag) {
+				if(el && el.type == ElementType.point && el.style == PointStyle.action) {
+					el.style = PointStyle.normal;
+					existingRefs.push(el);
+					existingMarks.push(el.pt);
+					break;
+				}
+			}
+		}
+	}
+	useStore.setState({
+		existingRefs, existingMarks, existingLines, // Add the existing marks and lines to the store
+		solutions: [], // Empty out the list of solutions so the user notices that the input mark/line is solved
+	});
+
+	// Recreate the database
+	resetWorker(useDB.getState());
+}
+
+function AddButton({ data }: AddButtonProps) {
+	const { t } = useTranslation();
+	return (
+		<button className="btn btn-success py-0" onClick={() => addAsExistingCreases(data)}>
+			<i className="fa-solid fa-circle-plus"></i>
+			<span className="d-inline-block capitalize m-1">&nbsp;{t("phrase.addAsExistingRefs")}</span>
+		</button>
+	);
+}
+
 export function SolutionComponent({ data, show, onSelect }: SolutionComponentProps) {
 	const { t } = useTranslation();
 	const settings = useSettings();
@@ -49,12 +106,26 @@ export function SolutionComponent({ data, show, onSelect }: SolutionComponentPro
 	return (
 		<div className={"card mt-3 " + (show ? "" : "d-sm-none")} style={{ overflow: "hidden" }}>
 			<div className="card-header d-none d-sm-block">
-				<Exact err={data.err} />
-				<span className="d-inline-block capitalize">{t("phrase.solution")} {solution},</span> <span className="d-inline-block">{t("phrase.error")} {err},</span> <span className="d-inline-block">rank {data.rank}</span>
+				<div className="row align-items-center gx-1 gy-2">
+					<div className="col-auto">
+						<Exact err={data.err} />&nbsp;<span className="capitalize">{t("phrase.solution")} {solution},</span>
+					</div>
+					<div className="col-auto">{t("phrase.error")} {err},</div>
+					<div className="col-auto">rank {data.rank}</div>
+					<div className="col d-none d-md-block text-end" style={{ flex: "1 0 max-content" }}>
+						<AddButton data={data} />
+					</div>
+				</div>
 			</div>
 			{show ?
 				<div ref={ref} className="card-header d-sm-none text-bg-primary">
-					<span className="d-inline-block capitalize">{t("phrase.solution")} {solution},</span> <span className="d-inline-block">{t("phrase.error")} {err},</span> <span className="d-inline-block">rank {data.rank}</span>
+					<div className="row align-items-center gx-1">
+						<div className="col-auto">
+							<Exact err={data.err} />&nbsp;<span className="capitalize">{t("phrase.solution")} {solution},</span>
+						</div>
+						<div className="col-auto">{t("phrase.error")} {err},</div>
+						<div className="col-auto">rank {data.rank}</div>
+					</div>
 				</div> :
 				<div ref={ref} className="card-header d-sm-none" onClick={handleSelect} style={{ cursor: "pointer" }}>
 					<div className="row gx-0 justify-content-center align-items-top">
@@ -98,6 +169,9 @@ export function SolutionComponent({ data, show, onSelect }: SolutionComponentPro
 							</div>
 						</div>
 					)}
+				</div>
+				<div className="text-end d-md-none mt-2">
+					<AddButton data={data} />
 				</div>
 			</div>
 		</div>