+++ src/core/class/refLine/refLineL2LP2L.cpp
@@ -121,15 +121,14 @@ JsonObject RefLine_L2L_P2L::Serialize() const {
 	JsonObject step;
 	step.add("axiom", 7);
 	rl2->PutName("l1", step);
+	rm1->PutName("p0", step);
+	rl1->PutName("l0", step);
 	switch(mWhoMoves) {
 	case WHOMOVES_P1:
-		rm1->PutName("p0", step);
-		rl1->PutName("l0", step);
+		step.add("order", "p0,l0");
 		break;
-
 	case WHOMOVES_L1:
-		rl1->PutName("p0", step);
-		rm1->PutName("l0", step);
+		step.add("order", "l0,p0");
 		break;
 	};
 	PutName("x", step);